schepperbackend/functions/requirements.txt
firebase_functions~=0.2.0
nordigen==1.3.2
qrcode==7.3.1
pypdf2==3.0.1
fpdf2==2.7.6
pytz==2022.1
pydantic==2.5.3
email-validator==2.1.0.post1